Explicitly pass implicit parameters for some type class instances for Promise, as a workaround for https://lampsvn.epfl.ch/trac/scala/ticket/2316.
I had to tweak Pretty.prettyArgs to compile. This is in the scalacheck sources, so it's not checked in.

Only compiles with scalac from http://github.com/paulp/scala/tree/scalacheck as a workaround for https://lampsvn.epfl.ch/trac/scala/ticket/2201.

Removed implementation of Traverse[Array] and Pure[Array]. As far as I understand, we would need to change, for example:

trait Pure[+P[_]] {
  def pure[A](a: => A): P[A]
}

to:

trait Pure[+P[_]] {
  def pure[A: ClassManifest](a: => A): P[A]
}

This effectively adds an implicit parameter of type ClassManifest[A] to the signature. The change would need to be propagated up the call chain.


git-svn-id: http://scalaz.googlecode.com/svn@828 ea79b6a6-c355-11dd-8bad-cde060d8fd3c